# capnp: Refactor Arenas