@matheusd 2024-08-29
Rebased and updated to fix the final data races. Also updated the stats against latest main branch.